add squad dataset and parsing script, log_reg changes to save models
